add  search  function

+++ webService/src/main/java/com/uf/store/dao/mysql/ProductRepository.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
@@ -13,5 +12,5 @@ import com.uf.store.dao.mysql.po.Product;
 
 public interface ProductRepository  extends PagingAndSortingRepository<Product, Long> {
 	@Query("select p from Product p  where p.id in :ids ")
-	public Page<Product> findByIdIn(@Param("ids")List<Integer> ids,Pageable pageable);
+	public Page<Product> findByIdIn(@Param("ids")List<Long> ids,Pageable pageable);
 }
